Reflex association: Subject/subsubjects
